Python/refactor/requirements.txt
python-dotenv
python-telegram-bot
websockets
selenium
requests
